add MembershipTab
