+++ src/components/ui/member-details.tsx
@@ -10,18 +10,12 @@ import {
   parseYear,
   isAdult,
   isValidISODate,
-  getCustomDate,
   hasExpired,
-  getDateWeekLater,
-  getDateMonthsLater,
   serializeForUpdate,
 } from "@/lib/utils";
 import { useTranslation } from "react-i18next";
 import { useState } from "react";
-import { Ban, RefreshCcw } from "lucide-react";
-import { Popover, PopoverContent, PopoverTrigger } from "./popover";
 import { Textarea } from "./textarea";
-
 import {
   Form,
   FormControl,
@@ -30,7 +24,6 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { Input } from "./input";
 import {
   Sheet,
   SheetContent,
@@ -38,7 +31,6 @@ import {
   SheetTitle,
   SheetTrigger,
 } from "@/components/ui/sheet";
-
 import {
   Tooltip,
   TooltipContent,
@@ -50,6 +42,7 @@ import { useWindowSize } from "@/hooks/use-window-size";
 import { UpdateMutation } from "@/hooks/use-table-mutations";
 import { Member } from "@/types";
 import { PersonalTab } from "./personal-tab";
+import { MembershipTab } from "./membership-tab";
 
 export function MemberDetails({
   row,
@@ -66,7 +59,6 @@ export function MemberDetails({
   const [day, setDay] = useState(parseDay(row.birthDate));
   const [month, setMonth] = useState(parseMonth(row.birthDate));
   const [year, setYear] = useState(parseYear(row.birthDate));
-  const focusDelay = 50;
   const isMobile = useWindowSize();
 
   const formSchema = z.object({
@@ -122,7 +114,6 @@ export function MemberDetails({
 
   const isSuspended: boolean = Boolean(form.watch("suspendedTill"));
   const isExpired = hasExpired(new Date(form.watch("expirationDate")));
-  const isRenewAllowed = !isSuspended && isExpired;
   const isActive = !isSuspended && !isExpired;
   const [open, setOpen] = useState(false);
 
@@ -181,253 +172,12 @@ export function MemberDetails({
                   />
                 </TabsContent>
                 <TabsContent value="membership">
-                  <div className="flex flex-col space-y-8">
-                    <FormField
-                      name="registrationDate"
-                      render={() => (
-                        <FormItem>
-                          <FormLabel>{t("memberDetails.registered")}</FormLabel>
-                          <FormControl>
-                            <Input
-                              disabled={true}
-                              value={getCustomDate(row.registrationDate)}
-                            />
-                          </FormControl>
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name="expirationDate"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>
-                            {isExpired
-                              ? t("memberDetails.expired")
-                              : t("memberDetails.expires")}
-                          </FormLabel>
-                          <FormControl>
-                            <Input
-                              disabled={true}
-                              placeholder={t("memberDetails.notSuspended")}
-                              {...field}
-                              value={getCustomDate(
-                                form.watch("expirationDate"),
-                              )}
-                            />
-                          </FormControl>
-                        </FormItem>
-                      )}
-                    />
-                    <div className="flex">
-                      <Button
-                        disabled={
-                          form.formState.isSubmitting || !isRenewAllowed
-                        }
-                        type="button"
-                        variant="active"
-                        className="self-start"
-                        onClick={() =>
-                          form.setValue(
-                            "expirationDate",
-                            getDateMonthsLater(12),
-                            {
-                              shouldDirty: true,
-                            },
-                          )
-                        }
-                      >
-                        <RefreshCcw className={`w-5 mr-3`} />
-                        {t("memberDetails.renew")}
-                      </Button>
-                    </div>
-                    <div className="flex flex-col gap-4">
-                      <FormField
-                        control={form.control}
-                        name="suspendedTill"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>
-                              {t("memberDetails.suspended")}
-                            </FormLabel>
-                            <FormControl>
-                              <Input
-                                disabled={true}
-                                placeholder={t("memberDetails.notSuspended")}
-                                {...field}
-                                value={
-                                  isSuspended
-                                    ? getCustomDate(form.watch("suspendedTill"))
-                                    : ""
-                                }
-                              />
-                            </FormControl>
-                          </FormItem>
-                        )}
-                      />
-                      <div className="flex flex-col space-y-6">
-                        <FormField
-                          control={form.control}
-                          name="measure"
-                          render={({ field }) => (
-                            <FormItem>
-                              <FormLabel>
-                                {t("memberDetails.suspensionLabel")}
-                              </FormLabel>
-                              <FormControl>
-                                <Textarea
-                                  disabled={!isSuspended}
-                                  className="resize-none"
-                                  {...field}
-                                />
-                              </FormControl>
-                              <FormMessage />
-                            </FormItem>
-                          )}
-                        />
-                        {isSuspended ? (
-                          <Button
-                            disabled={form.formState.isSubmitting}
-                            type="button"
-                            variant="suspended"
-                            className="self-start"
-                            onClick={() => {
-                              form.setValue("suspendedTill", "", {
-                                shouldDirty: true,
-                              });
-                              form.setValue("measure", "", {
-                                shouldDirty: true,
-                              });
-                            }}
-                          >
-                            <Ban className={"w-5 mr-3"} />
-                            {t("memberDetails.resume")}
-                          </Button>
-                        ) : (
-                          <Popover>
-                            <PopoverTrigger asChild>
-                              <Button
-                                disabled={
-                                  form.formState.isSubmitting || isSuspended
-                                }
-                                type="button"
-                                variant="suspended"
-                                className="self-start"
-                              >
-                                <Ban className={"w-5 mr-3"} />
-                                {t("memberDetails.suspend")}
-                              </Button>
-                            </PopoverTrigger>
-                            <PopoverContent>
-                              <ul className="space-y-2">
-                                <li>
-                                  <Button
-                                    size="sm"
-                                    variant="suspended"
-                                    onClick={() => {
-                                      form.setValue(
-                                        "suspendedTill",
-                                        getDateWeekLater(),
-                                        {
-                                          shouldDirty: true,
-                                        },
-                                      );
-                                      setTimeout(() => {
-                                        form.setFocus("measure");
-                                      }, focusDelay);
-                                    }}
-                                  >
-                                    {t("durations.week", { count: 1 })}
-                                  </Button>
-                                </li>
-                                <li>
-                                  <Button
-                                    size="sm"
-                                    variant="suspended"
-                                    onClick={() => {
-                                      form.setValue(
-                                        "suspendedTill",
-                                        getDateMonthsLater(1),
-                                        {
-                                          shouldDirty: true,
-                                        },
-                                      );
-                                      setTimeout(() => {
-                                        form.setFocus("measure");
-                                      }, focusDelay);
-                                    }}
-                                  >
-                                    {t("durations.month", { count: 1 })}
-                                  </Button>
-                                </li>
-                                <li>
-                                  <Button
-                                    size="sm"
-                                    variant="suspended"
-                                    onClick={() => {
-                                      form.setValue(
-                                        "suspendedTill",
-                                        getDateMonthsLater(3),
-                                        {
-                                          shouldDirty: true,
-                                        },
-                                      );
-                                      setTimeout(() => {
-                                        form.setFocus("measure");
-                                      }, focusDelay);
-                                    }}
-                                  >
-                                    {t("durations.month", { count: 3 })}
-                                  </Button>
-                                </li>
-                                <li>
-                                  <Button
-                                    size="sm"
-                                    variant="suspended"
-                                    onClick={() => {
-                                      form.setValue(
-                                        "suspendedTill",
-                                        getDateMonthsLater(6),
-                                        {
-                                          shouldDirty: true,
-                                        },
-                                      );
-                                      setTimeout(() => {
-                                        form.setFocus("measure");
-                                      }, focusDelay);
-                                    }}
-                                  >
-                                    {t("durations.month", { count: 6 })}
-                                  </Button>
-                                </li>
-                                <li>
-                                  <Button
-                                    size="sm"
-                                    variant="suspended"
-                                    onClick={() => {
-                                      form.setValue(
-                                        "suspendedTill",
-                                        getDateMonthsLater(12),
-                                        {
-                                          shouldDirty: true,
-                                        },
-                                      );
-                                      setTimeout(() => {
-                                        form.setFocus("measure");
-                                      }, focusDelay);
-                                    }}
-                                  >
-                                    {t("durations.year", { count: 1 })}
-                                  </Button>
-                                </li>
-                              </ul>
-                            </PopoverContent>
-                          </Popover>
-                        )}
-                      </div>
-                    </div>
-                  </div>
+                  <MembershipTab
+                    form={form}
+                    row={row}
+                    isExpired={isExpired}
+                    isSuspended={isSuspended}
+                  />
                 </TabsContent>
                 <TabsContent value="note">
                   <FormField
